Extrais toute la logique de mise en avant et découpe du rideau dans la modale de sous étape
...pour permettre par la suite de piloter l'intégralité
